Integrate Update into Index
