+++ meilisearch/index.py
@@ -1,12 +1,11 @@
 from meilisearch._httprequests import HttpRequests
-from meilisearch.update import Update
 from meilisearch.document import Document
 from meilisearch.search import Search
 from meilisearch.stat import Stat
 from meilisearch.setting import Setting
 
 # pylint: disable=too-many-ancestors
-class Index(Update, Document, Search, Stat, Setting):
+class Index(Document, Search, Stat, Setting):
     """
     Indexes routes wrapper
 
@@ -19,7 +18,40 @@ class Index(Update, Document, Search, Stat, Setting):
         Index url path
     """
     index_path = 'indexes'
+    update_path = 'updates'
     uid = ""
+
+    class Update:
+        """
+        Update routes wrapper
+
+        Index's parent that gives access to all the update methods of MeiliSearch.
+        https://docs.meilisearch.com/references/updates.html#get-an-update-status
+
+        Attributes
+        ----------
+        update_path:
+            Update url path
+        """
+
+
+        def __init__(self, parent_path, config, uid=None, name=None):
+            """
+            Parameters
+            ----------
+            config : Config
+                Config object containing permission and location of MeiliSearch
+            name: str
+                Name of the index on which to perform the index actions.
+            uid: str
+                Uid of the index on which to perform the index actions.
+            index_path: str
+                Index url path
+            """
+            self.config = config
+            self.name = name
+            self.uid = uid
+            self.index_path = parent_path
 
     def __init__(self, config, uid):
         """
@@ -32,7 +64,6 @@ class Index(Update, Document, Search, Stat, Setting):
         index_path: str
             Index url path
         """
-        Update.__init__(self, Index.index_path, config, uid)
         Search.__init__(self, Index.index_path, config, uid)
         Document.__init__(self, Index.index_path, config, uid)
         Stat.__init__(self, Index.index_path, config, uid)
@@ -154,3 +185,42 @@ class Index(Update, Document, Search, Stat, Setting):
         if uid is not None:
             return Index(config, uid=uid)
         raise Exception('Uid is needed to find index')
+
+    def get_all_update_status(self):
+        """Get all update status from MeiliSearch
+
+        Returns
+        ----------
+        update: `list`
+            List of all enqueued and processed actions of the index.
+        """
+        return HttpRequests.get(
+            self.config,
+            '{}/{}/{}'.format(
+                self.index_path,
+                self.uid,
+                self.update_path
+            )
+        )
+
+    def get_update_status(self, update_id):
+        """Get one update from MeiliSearch
+
+        Parameters
+        ----------
+        update_id: int
+            identifier of the update to retieve
+        Returns
+        ----------
+        update: `list`
+            List of all enqueued and processed actions of the index.
+        """
+        return HttpRequests.get(
+            self.config,
+            '{}/{}/{}/{}'.format(
+                self.index_path,
+                self.uid,
+                self.update_path,
+                update_id
+            )
+        )
